requirements.txt
numpy>=1.24.0
pandas>=2.0.0
scikit-learn>=1.2.0
torch>=2.0.0
matplotlib>=3.7.0
seaborn>=0.12.0
tqdm>=4.65.0
python-dateutil>=2.8.2
PyYAML>=6.0
pytest>=7.3.1 